+++ src/test/java/io/auklet/TestingTools.java
@@ -1,17 +1,19 @@
 package io.auklet;
 
 import io.auklet.config.DeviceAuth;
-import io.auklet.core.AukletDaemonExecutor;
 import io.auklet.core.DataUsageMonitor;
 import io.auklet.platform.JavaPlatform;
-import io.auklet.util.ThreadUtil;
 import mjson.Json;
-import okhttp3.Request;
-import okhttp3.Response;
+import okhttp3.*;
 
 import java.io.File;
-import java.util.concurrent.TimeUnit;
+import java.io.IOException;
+import java.net.URISyntaxException;
+import java.nio.file.Files;
+import java.nio.file.Paths;
 
+import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.BDDMockito.given;
 import static org.mockito.Mockito.mock;
 
@@ -53,7 +55,7 @@ public class TestingTools {
             .set("brokers", "0.0.0.0")
             .set("port", "0000");
 
-    protected Auklet aukletConstructor() throws AukletException {
+    protected Auklet aukletConstructor() throws AukletException, IOException, URISyntaxException {
         Auklet mocked = mock(Auklet.class);
         given(mocked.getAppId()).willReturn("0123456789101112");
         given(mocked.getDeviceAuth()).willReturn(new DeviceAuth());
@@ -63,6 +65,17 @@ public class TestingTools {
         given(mocked.getConfigDir()).willReturn(new File(".auklet").getAbsoluteFile());
         given(mocked.getMqttThreads()).willReturn(2);
         given(mocked.getMacHash()).willReturn("");
+
+        given(mocked.doApiRequest(any(Request.Builder.class), anyString())).willReturn(
+                new Response.Builder()
+                        .request(new Request.Builder().url("https://api.auklet.io").method("get", null).header("Authorization", "JWT " + "123").build())
+                        .addHeader("Authorization", "JWT " + "123")
+                        .body(ResponseBody.create(MediaType.parse("application/json"),
+                                new String(Files.readAllBytes(Paths.get(getClass().getClassLoader().getResource("response.json").toURI())))))
+                        .code(201)
+                        .protocol(Protocol.HTTP_2)
+                        .message("This worked")
+                        .build());
         return mocked;
     }
 }
