+++ src/test/java/io/auklet/misc/JsonUtilTest.java
@@ -1,0 +1,13 @@
+package io.auklet.misc;
+
+import io.auklet.util.JsonUtil;
+import org.junit.jupiter.api.Test;
+
+import static org.junit.jupiter.api.Assertions.assertNotEquals;
+
+class JsonUtilTest {
+    @Test
+    void testReadJson() {
+        assertNotEquals("{}", JsonUtil.readJson("{1:1}"));
+    }
+}
