+++ src/test/java/io/auklet/sink/AbstractSinkTest.java
@@ -8,20 +8,19 @@ import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.TestInstance;
 
-import java.lang.reflect.InvocationTargetException;
+import java.io.IOException;
+import java.net.URISyntaxException;
 
 import static junit.framework.TestCase.assertNotNull;
 
 @TestInstance(TestInstance.Lifecycle.PER_CLASS)
 class AbstractSinkTest extends TestingTools {
     private AbstractSink abstractSink;
-    @BeforeAll void setup() throws AukletException {
+    @BeforeAll void setup() throws AukletException, IOException, URISyntaxException {
         final Auklet auklet = aukletConstructor();
         abstractSink = new AbstractSink() {
             @Override
-            protected void write(@NonNull byte[] bytes) throws AukletException {
-
-            }
+            protected void write(@NonNull byte[] bytes) throws AukletException { }
 
             @Override
             public void start(@NonNull Auklet agent) throws AukletException {
