+++ src/test/java/io/auklet/sink/AukletIoSinkTest.java
@@ -7,6 +7,9 @@ import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.TestInstance;
 
+import java.io.IOException;
+import java.net.URISyntaxException;
+
 import static org.junit.jupiter.api.Assertions.assertFalse;
 
 @TestInstance(TestInstance.Lifecycle.PER_CLASS)
@@ -14,7 +17,7 @@ class AukletIoSinkTest extends TestingTools {
     private AukletIoSink aukletIoSink;
     private Auklet auklet;
 
-    @BeforeAll void setup() throws AukletException, RuntimeException {
+    @BeforeAll void setup() throws AukletException, IOException, URISyntaxException {
         auklet = aukletConstructor();
         aukletIoSink = new AukletIoSink();
