+++ src/test/resources/response.json
@@ -1,0 +1,21 @@
+{
+  "client_id": "client_id_value",
+  "organization": "organization_value",
+  "client_password": "client_password_value",
+  "type": "articles",
+  "id": "id_value",
+  "data": [{
+
+    "attributes": {
+      "title": "JSON:API paints my bikeshed!",
+      "body": "The shortest article. Ever.",
+      "created": "2015-05-22T14:56:29.000Z",
+      "updated": "2015-05-22T14:56:28.000Z"
+    },
+    "relationships": {
+      "author": {
+        "data": {"id": "42", "type": "people"}
+      }
+    }
+  }]
+}
